fix up printf("%x")

+++ include/stdint.h
@@ -17,7 +17,7 @@ typedef int int32;
 typedef int int32_t;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long int uint64_t;
+typedef unsigned long long uint64_t;
 
 typedef char byte;
 typedef uint16 word;
